Add support for ISO time/date format in st.time_input / st.date_input (#9753)
## Describe your changes

Add support for ISO time/date format in `st.time_input` and
`st.date_input`. This improves a couple of scenarios:

1. If you want to pass a literal value, today you need to instantiate a
`datetime.time` or `datetime.date` object. With this change, you just
pass a string.
   
   ```py
   # TODAY
   import datetime
   st.date_input("Pick a date", datetime.date(2020, 01, 02))
   ```
   
   ...but after this PR you can simply do this:
   ```py
   # NEW
   st.date_input("Pick a date", "2020-01-02")
   ```
   
2. If you already have ISO time in a variable or dataframe, today you
need to convert it into a `datetime.*` object before passing it to
Streamlit. With this change, you just pass it as-is.
   
   ```py
   # TODAY
   st.date_input("Pick a date", datetime.date.fromisoformat(df["DATE"]))
   ```
   
   ...but after this PR you can simply do this:
   ```py
   # NEW
   st.date_input("Pick a date", df["DATE"])
   ```

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- **Explanation of why no additional tests are needed:** n/a
- **Unit Tests (JS and/or Python):** Python tests included
- **E2E Tests: None needed.** No impact on frontend.
- **Any manual testing needed?** No.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.