Make number input controls responsive to font size (#9807)
## Describe your changes

Fix the number input control width to be responsive to the root font
size (via rem) instead of using px size. This is required for the
upcoming changes for advanced theming.

## Testing Plan

- No logical or visual changes. Just uses rem instead of px. Will be
better tested once configuring the base font size is supported.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.