Fixes a bug where timezones with a non- x:00 utc offset would start at a non 00:00 timer for st.audio_input (#9791)
## Describe your changes

This PR makes it so that for ALL timezones, the `st.audio_input` timer
starts at 00:00. We used the built in date formatting to try to format
this timestamp in the most locale appropriate way, but one side effect
of doing this can be seen in #9631. This PR fixes said issue.

## GitHub Issue Link (if applicable)

This PR closes https://github.com/streamlit/streamlit/issues/9631

## Testing Plan

Added unit tests for this formatting utility that checks a few different
timezones, namely `Australia/Adelaide` which has a `x:30` utc offset.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.