Make alert padding responsive to font size (#9805)
## Describe your changes

Fix the alert padding to be responsive to the root font size instead of
using px size from basweb. This is required for the upcoming changes for
advanced theming.

## Testing Plan

- No logical or visual changes. Just uses rem instead of px. Will be
better tested once configuring the base font size is supported.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.