Make some component styles more robust (#9839)
## Describe your changes

This PR applies a couple of styles to components, which are also still
configured in global styles and are not planned to be removed. This is
intended to make the components a bit more robust/self-contained for
future changes.

## Testing Plan

- No logical or visual changes -> No tests required.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.