[chore] Add no-debugger eslint rule (#9868)
## Describe your changes

Straightforward eslint rule change to add
[no-debugger](https://eslint.org/docs/latest/rules/no-debugger). I
noticed this came up in a PR recently and was manually caught by a
reviewer. This makes it easier to prevent any accidental inclusions of
this.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.