Update audio buffer to handle channels better (#9885)
## Describe your changes

Firefox is providing two channels based on the microphone. I re-assessed
the WAV audio file and adjusted the conversion to work. It works now on
my machine, and we will do a little QA.

## GitHub Issue Link (if applicable)
closes #9799

## Testing Plan

- Rigorous manual testing

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.