Only show st.cache decorator spinners for cache misses and outermost function (#9956)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9951

This PR changes the cache spinner to show only for cache misses to avoid
sending unnecessary messages. Also, this PR automatically prevents
spinner elements to be added for nested cache functions, as having one
for the outermost cache function should be sufficient, because its
runtime is dependent on the inner logic.
This change prevents apps becoming slow / unresponsive in cases where
cached functions are nested inside of each other and are called very
often.

We might want to update the spinner logic in the future as the
`st.empty()` usage (also for chats) feels more hacky than it should be.
However, this PR basically enforces a best practice to disable showing
the spinner for nested cache functions which does not make sense anyways
and might come as a surprise to users who do not expect that the
`show_spinner=False` argument has to be set. It fixes the scenario
explained in https://github.com/streamlit/streamlit/issues/9951

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python): Added a unit test to make sure that for
nested cache functions, only a single spinner element (`empty` as its
surrogate) is created

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.