Move `showDevelopmentOptions` out of App.tsx  (#10009)
## Describe your changes

I encountered an issue with Fast Refresh that was caused by exporting
the `showDevelopmentOptions` function from `App.tsx`. This violates [the
vite
rule](https://github.com/vitejs/vite-plugin-react-swc#consistent-components-exports)
`consistent-components-exports`.

![Screenshot 2024-12-11 at 5 35
08 PM](https://github.com/user-attachments/assets/e6a28a9f-61de-4dce-9c07-5b03719614ba)


This PR moves this function into a separate non-component file since it
needs to be used in the component and also the test.

## GitHub Issue Link (if applicable)

## Testing Plan

Covered by existing tests. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.