Fix visualization of disabled feedback (#10030)
## Describe your changes

When a feedback widget was selected but also disabled, the selection was
not recognizable visually anymore. This PR here fixes that to keep a
visual indication also for disabled feedback widgets when selected.
The e2e test is updated to show the enabled and disabled widget state
together with a selection in taken snapshots.

## GitHub Issue Link (if applicable)

## Testing Plan

- E2E Tests
  - Updates to account for selected feedback widgets that are disabled in
snapshots

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.