[refactor] Utilize userEvent in tests (#10033)
## Describe your changes

- Migrates a vast majority of test files to utilize `userEvent` rather
than `fireEvent`.
- Before this, there were 200 usages of `fireEvent`. After, there are
38, representing an **81% decrease**.
- Any remaining usages of `fireEvent` require some extra changes to get
them to work with `userEvent`. I decided this was out of scope for this
PR since I wanted this to not touch any source, and be
non-invasive/repetitive as possible
- As a reminder for why `userEvent` is preferred over `fireEvent`, read
[this section on their
docs](https://testing-library.com/docs/user-event/intro/#differences-from-fireevent)

## GitHub Issue Link (if applicable)

## Testing Plan

- This updates vitest tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.